extract LoadPatches to separate package
